// types.ts
export interface Client {
  id: string;
  ruc: string;
  name: string;
  email: string;
  address: string;
  phone: string;
  type: 'CLIENTE' | 'PROVEEDOR' | 'AMBOS';
}